popdat

Module od

This is a standalone pipeline for generating a Scenario, starting from origin-destination data (also called desire lines), which gives a count of commuters between two zones, breaking down by mode.

Structs

DesireLine

This describes some number of commuters living in some named zone, working in another (or the same zone), and commuting using some mode.

NormalDistribution

A normal distribution of Durations.

Options

Zone 🔒

Enums

IncludeZonePolicy

Only desire lines starting and ending in zones matching this policy will be used.

Functions

create_zones 🔒

disaggregate

Generates a scenario from aggregated origin/destination data (DesireLines). The input describes an exact number of people, who live in one zone and work in another (possibly the same) and commute using some mode. For each of them, we just need to pick a specific home and workplace from the zones, and use the Options to pick departure times. We’ll wind up creating people who just take two trips daily: home -> work -> home.