Module popdat::od

This is an alternative pipeline for generating a Scenario, starting from origin-destination data (also called desire lines), which gives a count of commuters between two zones, breaking down by mode.

Maybe someday, we'll merge the two approaches, and make the first generate DesireLines as an intermediate step.

Structs

DesireLine
This describes some number of commuters living in some named zone, working in another (or the same zone), and commuting using some mode.

NormalDistribution
A normal distribution of Durations.

Options

Zone

Functions

create_zones

disaggregate
TODO Describe. In particular, how are polygons partly or fully outside the map's boundary handled? TODO Add an options struct to specify AM/PM time distribution, lunch trips, etc.